remove executeScript from ADR
